projects/trend-analysis/python/requirements.txt
beautifulsoup4==4.8.2
celery==4.4.0
connexion==2.6.0
Flask==1.1.1
Flask-Cors==3.0.8
imbalanced-learn==0.6.2
mysql-connector==2.2.9
numpy==1.18.2
pandas==1.0.3
psutil==5.7.0
pycountry==19.8.18 
pycrypto==2.6.1
scikit-learn==0.22.2
sh==1.12.14
chardet==4.0.0
forex_python==1.6
Faker==8.10.3
